requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cssselect2==0.7.0
django-widget-tweaks==1.5.0
frozenlist==1.4.1
html5lib==1.1
idna==3.10
lightgbm==4.5.0
lxml==5.3.0
mkl-service==2.4.0
multidict==6.1.0
oscrypto==1.3.0
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==5.0.0
pypng==0.20220715.0
python-bidi==0.6.0
python-decouple==3.8
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
svglib==1.5.1
tinycss2==1.3.0
twilio==9.3.2
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.12.1
